requirements.txt
mcp>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
plotly>=5.15.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
torch
anthropic
python-docx
flask>=2.3.0
werkzeug>=2.3.0
PyPDF2>=3.0.0
docx2txt>=0.8